feat(users): post new user endpoint with auth
We let clerk validate the authenticated request and retrieve the user data (email, userid, username aka. handle) from clerk.
Clerk allows us to require a username on registration (must be configured in the dashboard) so we don't need to implement that.
If the user already exists in our db, we just return 409 for now.